Refactor get_suse_container_urls using data structure for each distri
I have approach this problem organizing the data into a hash. The data
structure can hold information explicitly, so it is more ease to handle.

Each <distri><version> hold three information
- released_registry
- totest_registry
- available_arch

I use anonumous functions for `registries` to provide flexibility. For instance
you can choose the different registry based on aarch. or it is possible to
modify the sub to return an array in the future.

I tried to group some of the versions as they provide commonalities but regular
hash accept scalar as a key.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>