Add function validate_script_output_retry
Allow running validate_script_output with given number of retries,
similar to script_retry. It's basically the same function but
giving the chance to run the command $retry times with $delay
waiting time between them. This might be useful for commands
that are not 100% stable due to network issues and we need to try
2 or 3 times until it works.